requirements.txt
Flask==3.0.2
Flask-Session==0.5.0
Authlib==1.3.1
python-dotenv==1.0.1
requests==2.32.3

gunicorn==22.0.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-keyvault-keys==4.8.0
azure-keyvault-certificates==4.8.0

# Dev/test tools
pytest==8.3.3
pytest-mock==3.14.0
cryptography==42.0.5
